Termino de compilar interfaz por linea

+++ client/src/game/game.cpp
@@ -21,7 +21,7 @@
 
 
 Game::Game(Connector &connector, uint8_t game_id, 
-            uint8_t player_id, uint8_t mapId)
+            uint32_t player_id, uint8_t mapId)
 :   connector(std::move(connector)), gameId(game_id), 
     playerId(player_id), mapId(mapId), isDead(true) {}
 
@@ -70,7 +70,7 @@ void Game::run(){
 void Game::stop(){
     std::unique_lock<std::mutex> l(this->mutex);
     this->changesAsk.close();
-    for (int i = 0; i < this->threads.size(); ++i){
+    for (int i = 0; i < (int)this->threads.size(); ++i){
         (*(this->threads[i])).stop();
         (*(this->threads[i])).join();
     }
